fix(deps): remove high severity vulnerability and add extra travis check on PRS
The only exception is `webpack-svgstore-plugin` that still points to some older package versions
with `low` severity vulnerabilities and apparently the repo is not active that [the actual fix](https://github.com/mrsum/webpack-svgstore-plugin/pull/172)
could be merged. Because of that I'm limiting the scope to just `high` severity issues.

Note: I decided to include `audit-ci` because `yarn` does not provide (yet) a way to filter
violations by severity level.

fix(deps): upgrade packages so that `yarn lint` works with the current babel-eslint