Declare lowlight as a peerDependency
The lowlight package used by the extension-code-block-lowlight
extension is a dependency that must be installed in the tiptap
framework and the application using tiptap.

Declaring lowlight as a peer dependency allows the application
to control the version of lowlight used by the framework and
the application. As a result, the application doesn’t distribute
two versions of lowlight.