Add undoable option to ParseRule
This is more of a POC than a true PR. Is this a change you would support?

The reason I'm interested in this is I want to convert pasted video links into embedded videos (or similar for other media). But it's possible that someone actually wants a link instead of an embedded video. This logic allows them to Mod-z and have the text they pasted show up again.