fix(docs): correct capitalization in signing header
Lowercases the signing header for webhook requests.

The request header for signing webhook requests is displayed with capitalizations in the docs. 

HTTP headers are case insensitive, but if you're in (for example) a NextJS app and look for a `X-Hocuspocus-Signature-256` header in a webhook request, it'll return undefined. However, looking for a `x-hocuspocus-signature-256` header will work.