Prevent selection deletions at the end of list items from joining lists
If starting a selection at the end of a list item, then pressing delete, the selected text wasn't being deleted, but instead the following list item was joined to the end.

Now the selected text will be deleted.

Joining will only occur if the selection spans multiple nodes or it's collapsed.