Fix unmounting vue-3 editor in test environments
I'm not sure of the exact cause, is `css` being set to an object by something that's not the right
type or is it lacking the right polyfills in vitest with `happy-dom` and `jsdom`. All I know is I'm
getting this error:

```
TypeError: _a.remove is not a function
 ❯ Editor.unmount ../../node_modules/@tiptap/core/src/Editor.ts:185:15
```

So it's being set to something, that does not have the remove method. So updating this line to make
sure it actually exists before calling it.

This fixes unmounting the editor without throwing errors in vue 3 inside vitest