fix: allow `undefined` as a value for the `default` attribute key
Previously it was legal to specify:

```js

    const CustomNode = Node.create({
      name: 'customNode',
      addAttributes() {
        return {
          optionalAttr: {
            // default is set explicitly to undefined, so it should be preserved
            default: undefined,
          },
        }
      },
    })
```

but with the change introduced by PR #7198, it is no longer allowed to have a default that is set to `undefined`, it can only be set to `null`. This fixes that by checking the attributes object to see whether the `default` key is in the object.
Which should be explicit enough that the user wants that value to be the default, no matter what it's value actually is.