Prevent Vue 2 EditorContent components from destroying Editor instances currently in use by other EditorContent components.
Without checking if this EditorContent instance is currently assigned to the `contentComponent` property, it's possible for the EditorContent instance to destroy an Editor that is in use by another EditorContent instance.  Before destroying the editor's view, it should make sure that it is the last EditorContent to use the editor.

I've tested the change myself and it seems to be fine.  In my case, I have a text editor that I want to quickly move from one component to another.  Without this, the first component ends up destroying the text editor after it's already been put to use by the second component.  With this, the second component is allowed to take the Editor from the first component without the second component destroying it on them.